No more 'Byebug.start' to get full backtraces!
Now byebug maintains the real stack size so it can display full
backtraces even if `byebug` is placed deep in the callstack.
Byebug.start is kept because people might still want to use it for
performance reasons. But it's no longer needed in order to get correct
callstack information. Commit includes:
* Required changes in the C extension
* Tests for the new feature
* Guide updated accordingly