Don't try be smarter than the guy who wrote this
A while ago I removed a hack which was preventing some assertions from
being swallowed and as a result, the testing suite has been broken for
a while. Fortunately, seems that no functionality (except for breakpoint
deletion which is how I discovered this) has been broken since. Reenable
the hack.