One failing test for issue #13
Thanks to https://bugs.ruby-lang.org/issues/9321 I've been able to
isolate the bug (or one at least one of them) affecting the 'next'
command. The proper fix, and whether it should be fixed in ruby trunk
or in byebug, is still being discussed. If it's not fixed in trunk,
how to approach this seems far from trivial.