Fix (finally) post_mortem mode
The change is ruby core that was breaking this was
[r43561](https://bugs.ruby-lang.org/projects/ruby-trunk/repository/revisions/43561).
rb_errinfo is not set anymore in TracePoint API events, so we were not
correctly setting the exception and context information into it. The
correct way to set the exception in the raise event is actually throught
the TracePoint API, provided by the raised_exception method.

Also, $! seemed to be nil in the at_exit hook. I think this is because
$! is not set from rb_errinfo if there's not rescue or ensure block
(like it happens in post_mortem mode). To solve this I use a class
variable to keep the raised_exception instead of relying on $! (this
change was actually made in 1cbf5d2).