Fix last post_mortem test
We don't need to rescue and reraise in debug_file test method, an ensure
clause is enough. Not doing this has the benefit that the last raised
exception is the exception set in the original raise event and not the
one set in the reraise" event, so we can make assertions on it.