Change semantics of finish
This is a backwards incompatible change that changes the meaning of the
`finish` command. `finish n` means now literally "finish _n_ frames from
current position and then stop", which makes more sense than the old
"run program until frame number _n_".

The `Context#stop_return` method has been removed from the API because
it was just added to support `byebug` statements at the end of blocks
and now `Context#step_out` can perfectly handle this.

Now the API now can be easily extended in the future to support
breakpoints at the end of method, which could be useful to inspect the return
value, for example.

pry-byebug will need to be upgraded to support this.