More resilient tests for stepping
Now line numbers are specified inline in the code to be evaluated. This
makes the tests more resilient, because a single addition or removal in
the file they belong to will not break them as it used to happen before.
All test should be eventually migrated to this new style.