Removed dependency on debugger-linecache
This should fix code reload which has been buggy for a long time. Source file
caching was implemented through debugger-linecache but there was also some
internal code in byebug trying to do the same thing. This commit moves
everything to the same place.

The new module Byebug::Filecache is a minimalistic version of the whole
debugger-linecache gem.