Fix intermittent test failure.
In some cases, the created breakpoint would actually have ID of 100, so
the breakpoint would actually exist and the expectation would fail. The
solution is to make sure breakpoint with ID 100 is actually deleted
before checking the command "info breakopints 100". Just adding "delete
100" would still not work for those cases because the breakpoint
wouldn't exist but it would be the only breakpoint and in that case the
error message is different: "No breakpoints.". So we also make sure
there's at least one breakpoint by creating another one ("break 38").