Don't allow catch command in control mode
Related to #92. Catch command is not allow anymore in control mode. It
was broken and I don't think it's worth fixing it. When loading the init
script, there's not an available context with associated
file/line/binding so we can not check whether the class to be catched is
defined or not. I can think of some workarounds but this is exactly how
"set post_mortem" is meant to be used so I'm just banning it.