Fix issue with class/b_call end/b_return events
This kind of events had the same callbacks as the "call" event, where a
method_id is assumed the be included in the TracePoint information.
However, this kind of events don't contain that kind of information.
This was the reason for: https://bugs.ruby-lang.org/issues/9937. The fix
consists on associating to those events the simpler callback for the c_call
event that only updates calculated callstack size information