Stop leaking state during tests
Previously we would instance a new processor and interface for every
test, load "convenient" settings for testing and don't care about the
final state of the debugger. This has proven to be prone to errors.

Now we reuse the same debugger instance for all tests and make sure we
leave the same state after every test. Also, we don't use any special
settings and just use byebug's defaults, so now the tests are testing
more real use cases.

This also led to several refactorings and some bug fixes (in the tests,
not in the actual code).