Include Gemfile.lock in source control
I don't really agree with the reasons not to do it stated in the popular
Yehuda Katz post ("Clarifying the Roles of the .gemspec and Gemfile")

* "It enforces precision that does not exist in the gem command, which
is used to install gems in practice" is outdated. I use `bundler` to
install development dependencies.
* "Even if the precision could be enforced, you wouldn't want it, since
it would prevent people from using your library with versions of its
dependencies that are different from the ones you used to develop the
gem". This is wrong, `Gemfile.lock` is ignored by rubygems. I don't even
ship it with the gem.