Big refactoring of byebug's internals. Fixes #140
The underlying command processor not being available to the command
itself complicated this issue. I could access a global list of commands,
but the command list is actually dependent on the debugger's mode.

Fixing this led to major refactorings in the whole debugger. Now every
command has the following methods available:

* `processor` gives access the the current command processor. This
method makes it so the `help` command can access the list of available
commands and their help messages.

* `context` gives access to the current context. It's the old @state
variable. The whole @state class has been removed because it was not
semantically different from the context itself.

* `frame` contains information about the current frame.