Fix bug in evaluation
When the user would enter an expression with a syntax error, the
exception would not be properly rescued and the user would lose the
command prompt.

This includes another change that fixes the semantics of the
`stack_on_error` settings. This setting is supposed to control the
verbosity of the error messages and not whether the exceptions are
actually rescued or not.