How to list if there's no source file
A common scenario of missing source files is debugging through dynamically generated code. In this case (and probably in any other case where source file is not to be found), it may be better to step through the source-less code and to resume debugging at the next available source line.
Anyway, just failing on missing source file crashes byebug on dynamically generated code.