Fix bug in `finish 0` command
Sometimes the `finish 0` command would fail to stop right before the
current frame would finish.

There was a bug in the counters (`step_out` needs to be increased also
when it's zero). Fixing that bug revealed the need to update the
`step_out` counter for every calling/returning event if we want `finish
0` to work in all cases. Since we are now considering c-return events
for stepping out, we want to change `byebug`'s entrypoint implementation
since it now needs to "step out" 3 events, not 2.