Fix #183
Apparently `rb_tracearg_event_flag` was not public API in 2.0. Instead
of trying to detect the type of event, just add separate handlers. This
is also better since we should print any return event information in the
case of the end of class definitions and in the previous version we were
informing that the return value was nil.