Allow setting the "verbose" flag at runtime
We used to provide this as a `byebug` setting but it was removed. While
I still don't see the value in providing it as a setting (it was added
to help troubleshooting TracePoint API bugs, but now it seems 99% bug
free), I find it useful to provide an easy way of enabling it.