Fix bug in evaluation
When evaluating stuff from the command prompt, I noticed that sometimes
we would get some incorrect information about having stopped at a
breakpoint. This would happen when a breakpoint is currently set in the
line where the evaluation is done, namely, if we've arrived to the
current prompt via "breakpoint" and not via "stepping".

Turns out that the file/line information "Binding#eval" uses is
inherited from the current position, so the TracePoint API event
generated for the line where `eval` is run would contain information
matching a breakpoint in the current line.

To fix it, we explicitly pass (special) information to eval to mark that
we are running it from `byebug`'s prompt.

While writing the test for this I found out a bug in some minitest custom
matchers, so I fixed that too.