Fix tests on Ruby 2.5
On Ruby 2.5, some location of
`rb_debug_inspector_backtrace_locations` may be changed.
This change affects the line number of current frame
when breakpoint is set to method. I think this change will not
affect byebug users because after this change the line number of
current frame is still inside of target method.
So, this commit adds branches to test cases to fix tests on Ruby 2.5.