Don't error out on warnings when compiling
This is the default for ruby releases, but it's not for the master
branch. Currently the master branch does not compile against clang with
this option turned on, so I'm disabling it. Even if it compiled, it
looks more useful to us to actually be able to test byebug even if ruby
compilation has some warnings.