Adds line # to make the filename clickable
Adding the exact line number of the byebug statement to the end of
the filename on the output makes the statement clickable in IDE's
and editors supporting this functionality. Currently clicking the
path goes to the top of the file. With the modification, clicking
takes the user to the exact byebug line in the file.