Remove old workaround (#581)
A while ago, we used to provide the same handler for `:call`, `:b_call`, and
`:class` events. At that time, we need to make this exception because
`:b_call` events don't include `method_id` at the top level.

Nowadays, this is part of the `:call` handler, while the other events
are handled separately. So the tracepoint information always include the
`method_id` and we can remove the workaround.