Workaround coverage reporting to codeclimate (#623)
Codeclimate's test-reporter does not yet support Github Actions. To
workaround that, we need to make sure we pass the generic `GIT_BRANCH`
and `GIT_COMMIT_SHA` environment variables to the command that generates
the reports in the format codeclimate understands.

Since this information is not available by default in Github Actions, we
do a bit of `git` manipulation to find it.