Improve reliability of a test
This test is coded to be run against a non-monkeypatched version of
`require`, because it expects that by jumping out a single level with
`finish`, we will end up inside user code. This is the case when the
test is run through `bundler`, because `bundler` doesn't need to
monkeypatch `require`, but it's not the case when `byebug` is run
without `bundler`, because in that case a single "jump out" will land
inside the monkeypatched `require` code provided by rubygems.

I always kind of expected the full test suite to be run with `bundler`,
but since all of the other test pass on a "bundler-free" environment,
I'll add this patch to make it fully green.