Disable minitest plugins
The latest version of simplecov comes with a minitest plugin in order to
play nice with minitest's `at_exit` hook. However, we don't use
Minitest's `at_exit` hook, so it actually breaks things for us, because
now we don't get `simplecov`'s `at_exit` hooks at all.

To workaround it, I disable minitest plugin loading so that the previous
behavior will be kept when upgrading simplecov.