Use a simpler Gemfile on Windows
Our Windows head CI has issues because the lockfile gets set to a platform
specific version of ffi, but that version does not support ruby-head
(3.0).

Since the ffi gem is only needed for linting (as a depdendency of mdl),
and we don't lint on Windows, let's use a simpler `Gemfile` not
including linting gems on Windows.