Add 'fiddle' gem as a runtime dependency
Fixes deivid-rodriguez/byebug#851

The following warning is printed to stderr when using byebug on Ruby
3.3.5:

> /home/david/.rbenv/versions/3.3.5/lib/ruby/3.3.0/reline.rb:9: warning:
> fiddle was loaded from the standard library, but will no longer be
> part of the default gems starting from Ruby 3.5.0.
>
> You can add fiddle to your Gemfile or gemspec to silence this warning.

This change aims to cause that warning no longer to appear, by
implementing one of the suggestions in the warning message, i.e. by
adding the `fiddle` gem to byebug's gemspec as a runtime dependency.