Prefer faraday to net/http for fetching ruby release info (#913)
net-http is having issues with the latest openssl version, unless the
openssl gem is updated to 3.3.1. However, net-http does not declare
openssl as a dependency, so we need to artificially add it to the
Gemfile in order to use the latest version.

faraday seems to just work though, so let's use that.